added permissions

+++ src/Plugin/Service/PluginService.php
@@ -5,8 +5,12 @@ declare(strict_types=1);
 namespace Forumify\Plugin\Service;
 
 use Composer\InstalledVersions;
+use Forumify\Core\Repository\PermissionRepository;
 use Forumify\Core\Repository\PluginRepository;
+use Forumify\ForumifyBundle;
+use Forumify\Plugin\AbstractForumifyPlugin;
 use Forumify\Plugin\Application\Service\PluginService as ApplicationPluginService;
+use Forumify\Plugin\Entity\Permission;
 use Forumify\Plugin\Entity\Plugin;
 use Symfony\Component\DependencyInjection\Attribute\Autowire;
 
@@ -23,6 +27,8 @@ class PluginService
         #[Autowire('%kernel.project_dir%')]
         private readonly string $rootDir,
         private readonly PluginRepository $pluginRepository,
+        private readonly PermissionRepository $permissionRepository,
+        private readonly ForumifyBundle $forumifyBundle,
     ) {
     }
 
@@ -34,6 +40,8 @@ class PluginService
             array_map(static fn (Plugin $plugin) => $plugin->getPackage(), $allPlugins),
             $allPlugins,
         );
+        $this->forumifyPermissions();
+
 
         $plugins = [];
         $installedPlugins = $this->getInstalledPlugins();
@@ -51,14 +59,52 @@ class PluginService
 
             $this->pluginRepository->save($plugin, false);
             $plugins[] = $plugin;
+
+            if (class_exists($pluginClass)) {
+                $pluginInstance = new $pluginClass();
+                if ($pluginInstance instanceof AbstractForumifyPlugin) {
+                    $permissions = $pluginInstance->getPluginPermissions()->getPermissions();
+
+                    $currentPermissions = $this->permissionRepository->findBy(['plugin' => $plugin]);
+
+                    foreach ($currentPermissions as $currentPermission) {
+                        if (!in_array($currentPermission->getPermission(), $permissions)) {
+                            $this->permissionRepository->remove($currentPermission);
+                        }
+                    }
+
+                    foreach ($permissions as $pluginPermission) {
+                        $existingPermission = $this->permissionRepository->findOneBy(['plugin' => $plugin, 'permission' => $pluginPermission]);
+                        if ($existingPermission === null) {
+                            $permission = new Permission();
+                            $permission->setPlugin($plugin);
+                            $permission->setPermission($pluginPermission);
+
+                            $this->permissionRepository->save($permission);
+                        }
+                    }
+                }
+            }
+
+            $permissions = $this->permissionRepository->findAll();
+
+            foreach ($permissions as $permission) {
+                $plugin = $permission->getPlugin();
+                if ($plugin === null || !$this->pluginRepository->find($plugin->getId())) {
+                    $this->permissionRepository->remove($permission);
+                }
+            }
+            $this->permissionRepository->flush();
         }
 
         foreach ($knownPlugins as $package => $plugin) {
+            if ($package === 'forumify/forumify-platform') {
+                continue;
+            }
             if (!isset($installedPlugins[$package])) {
                 $this->pluginRepository->remove($plugin, false);
             }
         }
-
         $this->pluginRepository->flush();
 
         $latestVersions = $this->getLatestVersions();
@@ -116,4 +162,36 @@ class PluginService
 
         return $foundPlugins;
     }
+
+    private function forumifyPermissions(): void
+    {
+        $plugin = $this->pluginRepository->findOneBy(['package' => 'forumify/forumify-platform']);
+
+        if (!$plugin) {
+            $forumify = new Plugin();
+            $forumify->setPluginClass('Forumify\ForumifyBundle');
+            $forumify->setPackage('forumify/forumify-platform');
+            $forumify->setLatestVersion('dev-master');
+            $forumify->setVersion('dev-master');
+            $forumify->setActive(false);
+
+
+            $this->pluginRepository->save($forumify);
+            $plugin = $forumify;
+        }
+
+        $permissions = $this->forumifyBundle->getPluginPermissions()->getPermissions();
+
+        foreach ($permissions as $forumifyPermission) {
+            $existingPermission = $this->permissionRepository->findOneBy(['plugin' => $plugin->getId(), 'permission' => $forumifyPermission]);
+            if ($existingPermission === null) {
+                $permission = new Permission();
+                $permission->setPlugin($plugin);
+                $permission->setPermission($forumifyPermission);
+
+                $this->permissionRepository->save($permission);
+            }
+        }
+    }
+
 }
